+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-None
+- Add `--with-config` flag to `kafkactl create topic`
 
 ## [v1.2.0] - 2023-04-30
 - Added support for username & password authentication (plaintext SASL)